Website updates #335 #336 (#338)
Removed the github Actions pkgdown workflow from this PR. #335 needs more work because the Action was failing when getting the env set up. Likely we need to update pkgdown in the renv.lock along with a bunch of other dependencies.

This PR updates the website with some acknowledgements and fixes the broken DESCRIPTION file where only one maintainer role can be allocated to authors.

I also testing updating igraph and rlang to match my local env, but that was causing the Ubuntu check to fail so I've dropped those and will test all the required package updates on a separate branch.