ren_dev, only install packages if on Unix like machine
The future package is used for asyc. And the only async feature we use is for stopping the optimization. Stop I believe only works well using MultiCore strategy. And the MultiCore strategy is only available on Unix-like machines. So for Windows, when the app is ran, the stop optimization button is hidden because the strategy is multisession.